backend/app/core/feedback_system.py
# A simplified example for DSPy integration
import dspy
from typing import List, Dict, Any

def process_feedback(feedback_entry: Dict[str, Any]):
    """
    A placeholder for a DSPy feedback loop.
    In a real implementation, this would use a DSPy program to optimize the agent.
    """
    print("Processing feedback for agent optimization...")
    # Example DSPy code (pseudo-code):
    # from dspy.primitives import Example
    # examples = [Example(question=entry["question"], solution=entry["solution"], rating=entry["rating"]) for entry in feedback_list]
    # program = AgenticProgram(Signature(...))
    # optimizer = dspy.teleprompt.BootstrapFewShot(...)
    # optimized_program = optimizer.compile(program, trainset=examples)
    # This would then update the agent's prompts or weights.
    print("DSPy optimization logic would be executed here.")